Rename 'destory' to 'destroy'
